Creating the server class, changing the calibration class for a function
